Remove the inline use and provide methods into the TranslationWrapper instead.
There might be other ways to fix this.
